Move Files implementations into a dedicated namespace to separate file collections from single files
