+++ bin/piqule-sync.php
@@ -5,17 +5,17 @@ declare(strict_types=1);
 
 use Haspadar\Piqule\Application\AnnouncedApplication;
 use Haspadar\Piqule\Application\FileApplication;
-use Haspadar\Piqule\File\CompositeFiles;
-use Haspadar\Piqule\File\DirectoryFiles;
 use Haspadar\Piqule\File\ExecutableFile;
 use Haspadar\Piqule\File\File;
 use Haspadar\Piqule\File\ForcedFile;
 use Haspadar\Piqule\File\InitialFile;
 use Haspadar\Piqule\File\InlineFile;
-use Haspadar\Piqule\File\ListedFiles;
-use Haspadar\Piqule\File\MappedFiles;
 use Haspadar\Piqule\File\Reaction\FileReactions;
 use Haspadar\Piqule\File\Reaction\ReportingFileReaction;
+use Haspadar\Piqule\Files\CompositeFiles;
+use Haspadar\Piqule\Files\DirectoryFiles;
+use Haspadar\Piqule\Files\ListedFiles;
+use Haspadar\Piqule\Files\MappedFiles;
 use Haspadar\Piqule\FileSystem\DiskFileSystem;
 use Haspadar\Piqule\FileSystem\DryRunFileSystem;
 use Haspadar\Piqule\FileSystem\Path;
