+++ src/Files/CompositeFiles.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\File;
+namespace Haspadar\Piqule\Files;
 
 use Override;
