+++ src/Files/Files.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\File;
+namespace Haspadar\Piqule\Files;
+
+use Haspadar\Piqule\File\File;
 
 interface Files
 {
