+++ tests/Unit/Application/FileApplicationTest.php
@@ -6,7 +6,7 @@ namespace Haspadar\Piqule\Tests\Unit\Application;
 
 use Haspadar\Piqule\Application\FileApplication;
 use Haspadar\Piqule\File\InlineFile;
-use Haspadar\Piqule\File\ListedFiles;
+use Haspadar\Piqule\Files\ListedFiles;
 use Haspadar\Piqule\FileSystem\InMemoryFileSystem;
 use Haspadar\Piqule\Tests\Unit\Fake\File\Reaction\FakeFileReaction;
 use PHPUnit\Framework\Attributes\Test;
