+++ tests/Unit/Files/ListedFilesTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Tests\Unit\File;
+namespace Haspadar\Piqule\Tests\Unit\Files;
 
 use Haspadar\Piqule\File\InlineFile;
-use Haspadar\Piqule\File\ListedFiles;
+use Haspadar\Piqule\Files\ListedFiles;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
