+++ tests/Unit/Files/MappedFilesTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Tests\Unit\File;
+namespace Haspadar\Piqule\Tests\Unit\Files;
 
 use Haspadar\Piqule\File\File;
 use Haspadar\Piqule\File\InlineFile;
-use Haspadar\Piqule\File\ListedFiles;
-use Haspadar\Piqule\File\MappedFiles;
+use Haspadar\Piqule\Files\ListedFiles;
+use Haspadar\Piqule\Files\MappedFiles;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
